Update project settings, refactor code, and enhance functionality
- Added file associations in VSCode settings for better C++ support.
- Updated C++ standard from C++17 to C++22 in CMakeLists.txt.
- Removed unnecessary executable files.
- Improved output formatting in decode.cpp.
- Added missing includes and refactored instruction handling in instruction.cpp and instructionex.cpp.
- Enhanced RAM memory management in ram.cpp and ram.h.
- Updated simulator VM to improve instruction handling and readability.
- Added .gitignore file to exclude unnecessary files from version control.

+++ app/main/decode.cpp
@@ -162,7 +162,7 @@ int main() {
     Decoder decoder;
     int code = 0x0064; // Example code
     Instruction instruction = decoder.decode(code);
-    std::cout << "Decoded instruction: " << static_cast<int>(instruction.getOpc()) << std::endl;
+    std::cout << "Decoded instruction: " << instruction.getOpc() << std::endl;
     return 0;
 }
